+++ src/snowflake/snowpark/_internal/utils.py
@@ -61,35 +61,11 @@ from snowflake.snowpark.version import VERSION as snowpark_version
 
 IS_V5_DRIVER: bool = connector_version[0] >= 5
 
-if IS_V5_DRIVER:
-    from snowflake.connector._common.extras import pandas  # noqa: F401
-    from snowflake.connector._common.extras import (
-        MissingOptionalDependency,
-        ModuleLikeObject,
-        pyarrow,
-        installed_pandas,
-        installed_pyarrow,
-    )
-else:
-    from snowflake.connector.options import pandas  # noqa: F401
-    from snowflake.connector.options import (
-        MissingOptionalDependency,
-        MissingPandas,
-        ModuleLikeObject,
-        pyarrow,
-        installed_pandas,
-    )
-
-    # connector.options (v4) never exported installed_pyarrow as its own name.
-    installed_pyarrow: bool = not isinstance(pyarrow, MissingOptionalDependency)
-
-
-def _missing_pandas() -> MissingOptionalDependency:
-    # v4 has no __init__ override (no-arg-subclass only); v5 deleted
-    # MissingPandas in favor of the positional-arg form.
-    if IS_V5_DRIVER:
-        return MissingOptionalDependency("pandas")
-    return MissingPandas()
+from snowflake.snowpark._internal.options import (  # noqa: E402
+    MissingOptionalDependency,
+    ModuleLikeObject,
+    installed_pandas,
+)
 
 
 if TYPE_CHECKING:
