+++ tests/integ/scala/test_update_delete_merge_suite.py
@@ -8,7 +8,7 @@ import datetime
 
 import pytest
 
-from snowflake.snowpark._internal.utils import installed_pandas
+from snowflake.snowpark._internal.options import installed_pandas
 from snowflake.snowpark import (
     DeleteResult,
     MergeResult,
